backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
google-genai==0.3.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
